NEW: ripeatlas component
This is the first commit for the ripeatlas component.
This component provide a wrapper for the protocol-ri around the
RIPE atlas-toolbox (https://github.com/pierdom/atlas-toolbox).

Each capability takes a command line style list of options as input,
and returns each line of the evoked atlas-toolbox script as a
different result